+++ ejb/src/main/java/it/polimi/telco/ejb/services/ProductService.java
@@ -1,19 +1,14 @@
 package it.polimi.telco.ejb.services;
 
 import it.polimi.telco.ejb.entities.Product;
-import it.polimi.telco.ejb.entities.Service;
-import it.polimi.telco.ejb.entities.ServicePackage;
-import it.polimi.telco.ejb.entities.Subscription;
 import it.polimi.telco.ejb.exceptions.NoProductFoundException;
-import it.polimi.telco.ejb.exceptions.NoServiceFoundException;
 
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceException;
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.util.List;
-import java.util.Set;
 
 @Stateless(name = "ProductServiceEJB")
 public class ProductService {
@@ -37,7 +32,7 @@ public class ProductService {
         return products;
     }
 
-    public Product prepareProduct(String productDescription, int monthlyFee){
+    public Product prepareProduct(String productDescription, BigDecimal monthlyFee){
 
         Product product = new Product();
